+++ gen/templates/provider.go
@@ -44,6 +44,9 @@ import (
 	"github.com/netascode/go-restconf"
 )
 
+var _ provider.Provider = &IosxeProvider{}
+var _ provider.ProviderWithActions = &IosxeProvider{}
+
 const (
 	YangPatch = false
 )
@@ -597,7 +600,7 @@ func (p *IosxeProvider) Configure(ctx context.Context, req provider.ConfigureReq
 	resp.ResourceData = &data
 }
 
-func (p *IosxeProvider) Resources(ctx context.Context) []func() resource.Resource {
+func (p *IosxeProvider) Resources(_ context.Context) []func() resource.Resource {
 	return []func() resource.Resource{
 		NewYangResource,
 		NewSaveConfigResource,
@@ -609,12 +612,18 @@ func (p *IosxeProvider) Resources(ctx context.Context) []func() resource.Resourc
 	}
 }
 
-func (p *IosxeProvider) DataSources(ctx context.Context) []func() datasource.DataSource {
+func (p *IosxeProvider) DataSources(_ context.Context) []func() datasource.DataSource {
 	return []func() datasource.DataSource{
 		NewYangDataSource,
 		{{- range .}}
 		New{{camelCase .Name}}DataSource,
 		{{- end}}
+	}
+}
+
+func (p *IosxeProvider) Actions(_ context.Context) []func() action.Action {
+	return []func() action.Action{
+		NewCommitAction,
 	}
 }