+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 
 - Add `icmp_echo_frequency` attribute to `iosxe_sla` resource and data source
+- Add `iosxe_commit` action
 
 ## 0.13.0